Conflicts:
	lib/edit/object.txt (amulet of magi)
	src/cmd3.c (refill_lamp moved to cmd-obj.c)
	src/load-old.c (removed)

+++ src/spells.h
@@ -92,7 +92,7 @@ extern bool destroy_doors_touch(void);
 extern bool sleep_monsters_touch(bool aware);
 extern bool curse_armor(void);
 extern bool curse_weapon(void);
-extern void brand_object(object_type *o_ptr, byte brand_type);
+extern void brand_object(object_type *o_ptr, int brand_type);
 extern void brand_weapon(void);
 extern bool brand_ammo(void);
 extern bool brand_bolts(void);
